pyproject.toml
[project]
name = "fmpxx"
version = "0.3.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "retry>=0.9.2",
    "agno>=1.7.5",
    "pydantic>=2.0.0",
    "google-genai>=1.26.0",
    "plotly>=6.2.0",
    "yfinance>=0.2.65",
]
